requirements.txt
fastapi
uvicorn[standard]
jinja2
mysql-connector-python
redis[hiredis]
python-dotenv
confluent-kafka
